requirements.txt
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.3.0
pillow>=8.0.0
scikit-image>=0.18.0
pandas>=1.3.0
pints>=0.5.0
